refactor: move components to new files
